# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.AuthProvider:
    properties:
      config:
        type: object
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.PublicAuthProvider:
    properties:
      authorize_path:
        description: AuthorizePath 企业微信类 OAuth 登录方式的发起地址，密码型认证方式为空。
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  router.authProviderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.AuthProvider'
        type: array
      total:
        type: integer
    type: object
  router.authProviderRequestDoc:
    properties:
      config:
        type: object
      enabled:
        example: true
        type: boolean
      name:
        example: AD/LDAP
        type: string
    type: object
...
  router.errorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  router.publicAuthProviderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.PublicAuthProvider'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: KVM 虚拟化资源管理控制台后端 API，提供认证、Agent 管理、运行态资源、刷新事件、任务、审计和告警接口。
  title: KVM Manager API
  version: "1.0"
paths:
...
  /api/auth/providers:
    get:
      description: 返回已启用的外部认证方式；本地账号登录始终可用。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.publicAuthProviderListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: 获取登录页可用认证方式
      tags:
      - auth
...
  /api/settings/auth-providers:
    get:
      description: 返回 AD/LDAP 等外部认证配置。LDAP bindPassword 不返回明文，已配置时返回 hasBindPassword
        标记。需要认证配置查看或管理权限。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.authProviderListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取认证配置
      tags:
      - settings
  /api/settings/auth-providers/{id}:
    put:
      consumes:
      - application/json
      description: 支持 AD/LDAP、企业微信直连（wecom）和企业微信统一认证中心（wecom_center）。显示名称不能为空；启用后登录页会显示对应认证方式。LDAP
        启用时必须提供服务器地址、端口、Base DN、用户过滤器、绑定 DN 和绑定密码，bindPassword 留空时保留已保存值；LDAPS 通常使用
        636 端口，StartTLS 通常使用 389 端口，二者不能同时启用。企业微信直连启用时必须提供企业 ID、应用 AgentId、应用 Secret
        和外部访问地址，secret 留空时保留已保存值，登录方式仅支持 qrcode（PC 扫码）和 inside（企微内置浏览器），开启调试模式（mock）时不校验
        AgentId 与 Secret 并使用模拟用户。统一认证中心启用时必须提供认证中心地址、应用标识和应用密钥，appSecret 留空时保留已保存值。关闭认证时允许保存空配置，用于清空已保存配置。外部认证用户必须先在用户配置中创建并启用后才可登录。
      parameters:
      - description: 认证配置 ID：ldap、wecom（企业微信直连）、wecom_center（企业微信统一认证中心）
        in: path
        name: id
        required: true
        type: string
      - description: 认证配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.authProviderRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthProvider'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新认证配置
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"